+++ constants/javaTypes.js
@@ -1,3 +1,4 @@
 export const EntityArmorStand = Java.type("net.minecraft.entity.item.EntityArmorStand");
 export const EntityItem = Java.type("net.minecraft.entity.item.EntityItem");
 export const EntityFishHook = Java.type("net.minecraft.entity.projectile.EntityFishHook");
+export const S2FPacketSetSlot = Java.type("net.minecraft.network.play.server.S2FPacketSetSlot");