+++ constants/sounds.js
@@ -1,5 +1,5 @@
 // Sounds now are instatiated right before playing, so that it does not cause Java error on switching the output device.
-export const NOTIFICATION_SOUND_SOURCE = { source: "notification.ogg" };
+export const NOTIFICATION_SOUND_SOURCE = { source: "notification.ogg", priority: true };
 export const TIMER_SOUND_SOURCE = { source: "notification-bell.ogg" };
 export const OH_MY_GOD_SOUND_SOURCE = { source: "oh-my-god.ogg", priority: true }; // Good items
 export const INSANE_SOUND_SOURCE = { source: "insane.ogg", priority: true }; // Insane RNG drops
@@ -7,15 +7,7 @@ export const AUGH_SOUND_SOURCE = { source: "augh.ogg", priority: true }; // Epic
 export const GOOFY_LAUGH_SOUND_SOURCE = { source: "goofy-laugh.ogg", priority: true }; // Rare pets
 export const WOW_SOUND_SOURCE = { source: "wow.ogg", priority: true };  // Legendary pets
 export const SHEESH_SOUND_SOURCE = { source: "sheesh.ogg", priority: true }; // Legendary yeti
-
-// export const NOTIFICATION_SOUND = new Sound({ source: "notification.ogg" });
-// export const TIMER_SOUND = new Sound({ source: "notification-bell.ogg" });
-// export const OH_MY_GOD_SOUND = new Sound({ source: "oh-my-god.ogg", priority: true }); // Good items
-// export const INSANE_SOUND = new Sound({ source: "insane.ogg", priority: true }); // Insane RNG drops
-// export const AUGH_SOUND = new Sound({ source: "augh.ogg", priority: true }); // Epic pets
-// export const GOOFY_LAUGH_SOUND = new Sound({ source: "goofy-laugh.ogg", priority: true }); // Rare pets
-// export const WOW_SOUND = new Sound({ source: "wow.ogg", priority: true });  // Legendary pets
-// export const SHEESH_SOUND = new Sound({ source: "sheesh.ogg", priority: true }); // Legendary yeti
+export const SAD_TROMBONE_SOUND_SOURCE = { source: "sad-trombone.ogg", priority: true };
 
 export const MEME_SOUND_MODE = 0;
 export const NORMAL_SOUND_MODE = 1;